# integration_test/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "rRPC API",
    "version": "0.1.0"
  },
  "paths": {
    "/rpc/test_empty": {
      "post": {
        "operationId": "TestEmpty",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestEmptyResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "examples": {
                  "validation": {
                    "value": {
                      "type": "validation",
                      "message": "validation error"
                    }
                  },
                  "input": {
                    "value": {
                      "type": "input",
                      "message": "input error"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "unauthorized",
                  "message": "unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "forbidden",
                  "message": "forbidden"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "custom",
                  "message": "error"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "not_implemented",
                  "message": "not implemented"
                }
              }
            }
          }
        }
      }
    },
    "/rpc/test_no_return": {
      "post": {
        "operationId": "TestNoReturn",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestNoReturnResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "examples": {
                  "validation": {
                    "value": {
                      "type": "validation",
                      "message": "validation error"
                    }
                  },
                  "input": {
                    "value": {
                      "type": "input",
                      "message": "input error"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "unauthorized",
                  "message": "unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "forbidden",
                  "message": "forbidden"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "custom",
                  "message": "error"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "not_implemented",
                  "message": "not implemented"
                }
              }
            }
          }
        }
      }
    },
    "/rpc/test_basic": {
      "post": {
        "operationId": "TestBasic",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestBasicParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestBasicResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "examples": {
                  "validation": {
                    "value": {
                      "type": "validation",
                      "message": "validation error"
                    }
                  },
                  "input": {
                    "value": {
                      "type": "input",
                      "message": "input error"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "unauthorized",
                  "message": "unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "forbidden",
                  "message": "forbidden"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "custom",
                  "message": "error"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "not_implemented",
                  "message": "not implemented"
                }
              }
            }
          }
        }
      }
    },
    "/rpc/test_list_map": {
      "post": {
        "operationId": "TestListMap",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestListMapParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestListMapResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "examples": {
                  "validation": {
                    "value": {
                      "type": "validation",
                      "message": "validation error"
                    }
                  },
                  "input": {
                    "value": {
                      "type": "input",
                      "message": "input error"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "unauthorized",
                  "message": "unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "forbidden",
                  "message": "forbidden"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "custom",
                  "message": "error"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "not_implemented",
                  "message": "not implemented"
                }
              }
            }
          }
        }
      }
    },
    "/rpc/test_optional": {
      "post": {
        "operationId": "TestOptional",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestOptionalParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestOptionalResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "examples": {
                  "validation": {
                    "value": {
                      "type": "validation",
                      "message": "validation error"
                    }
                  },
                  "input": {
                    "value": {
                      "type": "input",
                      "message": "input error"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "unauthorized",
                  "message": "unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "forbidden",
                  "message": "forbidden"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "custom",
                  "message": "error"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "not_implemented",
                  "message": "not implemented"
                }
              }
            }
          }
        }
      }
    },
    "/rpc/test_validation_error": {
      "post": {
        "operationId": "TestValidationError",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestValidationErrorParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestValidationErrorResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "examples": {
                  "validation": {
                    "value": {
                      "type": "validation",
                      "message": "validation error"
                    }
                  },
                  "input": {
                    "value": {
                      "type": "input",
                      "message": "input error"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "unauthorized",
                  "message": "unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "forbidden",
                  "message": "forbidden"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "custom",
                  "message": "error"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "not_implemented",
                  "message": "not implemented"
                }
              }
            }
          }
        }
      }
    },
    "/rpc/test_unauthorized_error": {
      "post": {
        "operationId": "TestUnauthorizedError",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestUnauthorizedErrorResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "examples": {
                  "validation": {
                    "value": {
                      "type": "validation",
                      "message": "validation error"
                    }
                  },
                  "input": {
                    "value": {
                      "type": "input",
                      "message": "input error"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "unauthorized",
                  "message": "unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "forbidden",
                  "message": "forbidden"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "custom",
                  "message": "error"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "not_implemented",
                  "message": "not implemented"
                }
              }
            }
          }
        }
      }
    },
    "/rpc/test_forbidden_error": {
      "post": {
        "operationId": "TestForbiddenError",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestForbiddenErrorResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "examples": {
                  "validation": {
                    "value": {
                      "type": "validation",
                      "message": "validation error"
                    }
                  },
                  "input": {
                    "value": {
                      "type": "input",
                      "message": "input error"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "unauthorized",
                  "message": "unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "forbidden",
                  "message": "forbidden"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "custom",
                  "message": "error"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RPCError"
                },
                "example": {
                  "type": "not_implemented",
                  "message": "not implemented"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TestEmptyResult": {
        "type": "object",
        "properties": {
          "empty": {"$ref":"#/components/schemas/EmptyModel"}
        }
      },
...
      "TestNoReturnResult": {
        "type": "object",
        "properties": {
        }
      },
      "TestBasicParams": {
        "type": "object",
        "properties": {
          "text": {"$ref":"#/components/schemas/TextModel"},
          "flag": {"type":"boolean"},
          "count": {"format":"int32","type":"integer"},
          "note": {"nullable":true,"type":"string"}
        },
        "required": ["text","flag","count"]
      },
      "TestBasicResult": {
        "type": "object",
        "properties": {
          "text": {"$ref":"#/components/schemas/TextModel"}
        }
      },
      "TestListMapParams": {
        "type": "object",
        "properties": {
          "texts": {"items":{"$ref":"#/components/schemas/TextModel"},"type":"array"},
          "flags": {"additionalProperties":{"type":"string"},"type":"object"}
        },
        "required": ["texts","flags"]
      },
      "TestListMapResult": {
        "type": "object",
        "properties": {
          "nested": {"$ref":"#/components/schemas/NestedModel"}
        }
      },
      "TestOptionalParams": {
        "type": "object",
        "properties": {
          "text": {"allOf":[{"$ref":"#/components/schemas/TextModel"}],"nullable":true},
          "flag": {"nullable":true,"type":"boolean"}
        }
      },
      "TestOptionalResult": {
        "type": "object",
        "properties": {
          "flags": {"$ref":"#/components/schemas/FlagsModel"}
        }
      },
      "TestValidationErrorParams": {
        "type": "object",
        "properties": {
          "text": {"$ref":"#/components/schemas/TextModel"}
        },
        "required": ["text"]
      },
      "TestValidationErrorResult": {
        "type": "object",
        "properties": {
          "text": {"$ref":"#/components/schemas/TextModel"}
        }
      },
...
      "TestUnauthorizedErrorResult": {
        "type": "object",
        "properties": {
          "empty": {"$ref":"#/components/schemas/EmptyModel"}
        }
      },
...
      "TestForbiddenErrorResult": {
        "type": "object",
        "properties": {
          "empty": {"$ref":"#/components/schemas/EmptyModel"}
        }
      },
...
      "RPCError": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "custom",
              "validation",
              "input",
              "unauthorized",
              "forbidden",
              "not_implemented"
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "required": ["type", "message"]
      }
    }
  }
}